// src/main/generated/typescript-helpers.tsx
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: any;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
};

export type ActiveCalendarInput = {
  key: Scalars['String'];
};

export type Area = {
  __typename?: 'Area';
  createdAt?: Maybe<Scalars['DateTime']>;
  deleted?: Maybe<Scalars['Boolean']>;
  deletedAt?: Maybe<Scalars['DateTime']>;
  description?: Maybe<Scalars['String']>;
  emoji?: Maybe<Scalars['String']>;
  items?: Maybe<Array<Maybe<Item>>>;
  key: Scalars['String'];
  lastUpdatedAt?: Maybe<Scalars['DateTime']>;
  name?: Maybe<Scalars['String']>;
  projects?: Maybe<Array<Maybe<Project>>>;
  sortOrder?: Maybe<AreaOrder>;
};

export type AreaOrder = {
  __typename?: 'AreaOrder';
  areaKey: Scalars['String'];
  sortOrder: Scalars['Int'];
};
// ...
export type Item = {
  __typename?: 'Item';
  area?: Maybe<Area>;
  children?: Maybe<Array<Maybe<Item>>>;
  completed?: Maybe<Scalars['Boolean']>;
  completedAt?: Maybe<Scalars['DateTime']>;
  createdAt?: Maybe<Scalars['DateTime']>;
  deleted?: Maybe<Scalars['Boolean']>;
  deletedAt?: Maybe<Scalars['DateTime']>;
  dueAt?: Maybe<Scalars['DateTime']>;
  key: Scalars['String'];
  label?: Maybe<Label>;
  lastUpdatedAt?: Maybe<Scalars['DateTime']>;
  parent?: Maybe<Item>;
  project?: Maybe<Project>;
  reminders?: Maybe<Array<Maybe<Reminder>>>;
  repeat?: Maybe<Scalars['String']>;
  scheduledAt?: Maybe<Scalars['DateTime']>;
  sortOrders?: Maybe<Array<Maybe<ItemOrder>>>;
  text?: Maybe<Scalars['String']>;
  type?: Maybe<Scalars['String']>;
};

export type ItemOrder = {
  __typename?: 'ItemOrder';
  componentKey: Scalars['String'];
  item: Item;
  sortOrder: Scalars['Int'];
};

export type Label = {
  __typename?: 'Label';
  colour?: Maybe<Scalars['String']>;
  key: Scalars['String'];
  name?: Maybe<Scalars['String']>;
};
// ...
export type Project = {
  __typename?: 'Project';
  area?: Maybe<Area>;
  createdAt?: Maybe<Scalars['DateTime']>;
  deleted?: Maybe<Scalars['Boolean']>;
  deletedAt?: Maybe<Scalars['DateTime']>;
  description?: Maybe<Scalars['String']>;
  emoji?: Maybe<Scalars['String']>;
  endAt?: Maybe<Scalars['DateTime']>;
  items?: Maybe<Array<Maybe<Item>>>;
  key: Scalars['String'];
  lastUpdatedAt?: Maybe<Scalars['DateTime']>;
  name: Scalars['String'];
  sortOrder?: Maybe<ProjectOrder>;
  startAt?: Maybe<Scalars['DateTime']>;
};

export type ProjectOrder = {
  __typename?: 'ProjectOrder';
  projectKey: Scalars['String'];
  sortOrder: Scalars['Int'];
};
// ...
export type Reminder = {
  __typename?: 'Reminder';
  createdAt?: Maybe<Scalars['DateTime']>;
  deleted?: Maybe<Scalars['Boolean']>;
  deletedAt?: Maybe<Scalars['DateTime']>;
  item?: Maybe<Item>;
  key: Scalars['String'];
  lastUpdatedAt?: Maybe<Scalars['DateTime']>;
  remindAt?: Maybe<Scalars['DateTime']>;
  text?: Maybe<Scalars['String']>;
};